requirements.txt
cachelib==0.6.0
certifi==2020.6.20
charset-normalizer==2.0.12
circuitbreaker==1.3.2
click==8.1.3
elastic-transport==8.1.2
elasticsearch==5.5.3
Flask==2.1.2
Flask-Dance==6.0.0
Flask-HTTPAuth==4.0.0
Flask-OAuthlib==0.9.6
Flask-SQLAlchemy==2.4.1
greenlet==1.1.2
gunicorn==20.1.0
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.2
kafka==1.3.5
kafka-python==2.0.2
MarkupSafe==2.1.1
oauthlib==2.1.0
PyJWT==1.7.1
pymongo==4.1.1
python-dotenv==0.20.0
requests==2.27.1
requests-oauthlib==1.1.0
SQLAlchemy==1.3.6
urllib3==1.26.9
URLObject==2.4.3
Werkzeug==2.0.0
